Add support for JAR file for Windows
