+++ src/main/java/duke/javafx/Launcher.java
@@ -1,12 +1,44 @@
 package duke.javafx;
 
+import java.io.IOException;
+
+import duke.Duke;
 import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
 
 /**
  * A launcher class to workaround classpath issues.
  */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
+public class Launcher extends Application {
+    private Duke duke = new Duke("data/duke.txt");
+
+    /**
+     * Sets scene and initialises duke.Duke.
+     *
+     * @param stage Stage to display.
+     */
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            fxmlLoader.<MainWindow>getController().setDuke(duke);
+            stage.setTitle("Duke");
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    /**
+     * Workaround for problems regarding Main extending Application.
+     */
+    public void launch() {
+        Application.launch();
     }
 }
