+++ src/main/java/duke/javafx/Main.java
@@ -1,37 +1,11 @@
 package duke.javafx;
 
-import java.io.IOException;
-
-import duke.Duke;
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.layout.AnchorPane;
-import javafx.stage.Stage;
-
 /**
  * A GUI for duke.Duke using FXML.
  */
-public class Main extends Application {
-    private Duke duke = new Duke("data/duke.txt");
-
-    /**
-     * Sets scene and initialises duke.Duke.
-     *
-     * @param stage Stage to display.
-     */
-    @Override
-    public void start(Stage stage) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
-            AnchorPane ap = fxmlLoader.load();
-            Scene scene = new Scene(ap);
-            stage.setScene(scene);
-            fxmlLoader.<MainWindow>getController().setDuke(duke);
-            stage.setTitle("Duke");
-            stage.show();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+public class Main {
+    public static void main(String[] args) {
+        new Launcher().launch();
     }
+
 }
